video-backend/requirements.txt
asgiref==3.8.1
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.17.18
awscrt==0.19.19
Babel==2.9.1
boto==2.49.0
boto3==1.35.90
botocore==1.35.90
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
Django==4.2.17
django-cors-headers==4.6.0
django-dynamodb-sessions==0.6
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.2
docutils==0.16
ec2-hibinit-agent==1.0.8
git-remote-codecommit==1.17
gpg==1.15.1
idna==2.10
ikp3db==1.4.2
Jinja2==2.11.3
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==3.0.2
mercurial==5.7.1
netifaces==0.10.6
oauthlib==3.0.2
packaging==21.3
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
psycopg2-binary==2.9.7
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
s3transfer==0.10.4
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
sqlparse==0.5.3
support-info==1.0
systemd-python==235
typing_extensions==4.12.2
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==2.3.7
